Split IndexerState out to it's own file.
