+++ src/snovault/elasticsearch/__init__.py
@@ -36,6 +36,7 @@ def includeme(config):
     config.include('.esstorage')
 
     config.include('.indexer')
+    config.include('.indexer_state')
     if asbool(settings.get('indexer')) and not PY2:
         config.include('.mpindexer')
